FEATURE: checkbox in typescript
